添加 NavbarContainer 组件
